# cmd/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Meetin APP API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/meeting/Service/multiple": {
            "put": {
                "description": "Update Multiple in UpdateService",
                "tags": [
                    "meeting Service"
                ],
                "summary": "Update Multiple in Service",
                "operationId": "ApiV1MettingAppServicesUpdateMultiple",
                "parameters": [
                    {
                        "description": "update params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_structures.ServiceEdit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api_structures.Service"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Create Multiple in CreateService",
                "tags": [
                    "meeting Service"
                ],
                "summary": "Create Multiple in Service",
                "operationId": "ApiV1MettingAppServiceCreateMultiple",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id uuid",
                        "name": "id",
                        "in": "path"
                    },
                    {
                        "description": "update params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_structures.ServiceFilter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api_structures.Service"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/meeting/Service/pagination": {
            "get": {
                "description": "Pagination GetServiceWithPagination",
                "tags": [
                    "meeting Service"
                ],
                "summary": "Pagination Service",
                "operationId": "ApiV1MettingAppServicePagination",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api_structures.Service"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/meeting/Service/search": {
            "get": {
                "description": "Get GetService",
                "tags": [
                    "meeting Service"
                ],
                "summary": "Show Service",
                "operationId": "ApiV1MettingAppServiceSearch",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "10",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "integrationId",
                        "description": "-id,name",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "[]",
                        "description": "[[col,eq,1],[col,eq,2],(missing and rule)[[col,eq,1],and|or,[col,eq,2]]]",
                        "name": "filters",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "Service,Service",
                        "description": "Service,Service",
                        "name": "preload",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "Service,Service",
                        "description": "Service,Service",
                        "name": "joins",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "integrationId,name,createdBy,createdAt,updatedAt,deletedAt",
                        "description": "integrationId,name,createdBy,createdAt,updatedAt,deletedAt",
                        "name": "fields",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api_structures.Service"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/meeting/Service/single": {
            "post": {
                "description": "Create in CreateService",
                "tags": [
                    "meeting Service"
                ],
                "summary": "Create in Service",
                "operationId": "ApiV1MettingAppServiceCreate",
                "parameters": [
                    {
                        "description": "update params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_structures.ServiceFilter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_structures.Service"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/meeting/Service/with-id/:id": {
            "get": {
                "description": "Get GetServiceById",
                "tags": [
                    "meeting Service"
                ],
                "summary": "Get Service by Unique IDs",
                "operationId": "ApiV1MettingAppServiceWithID",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Service,Service",
                        "description": "Service,Service",
                        "name": "preload",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "Service,Service",
                        "description": "Service,Service",
                        "name": "joins",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "integrationId,name,createdBy,createdAt,updatedAt,deletedAt",
                        "description": "integrationId,name,createdBy,createdAt,updatedAt,deletedAt",
                        "name": "fields",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api_structures.Service"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            },
            "put": {
                "description": "Update with id in UpdateService",
                "tags": [
                    "meeting Service"
                ],
                "summary": "Update with id in Service",
                "operationId": "ApiV1MettingAppServiceUpdateWithID",
                "parameters": [
                    {
                        "description": "update params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_structures.ServiceEdit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_structures.Service"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Delete with path params in DeleteService",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meeting Service"
                ],
                "summary": "Delete with path params in Service",
                "operationId": "ApiV1MettingAppServiceDelete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id uuid",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api_structures.Service": {
            "type": "object",
            "required": [
                "description",
                "price",
                "serviceName"
            ],
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "1960-11-10T04:48:09Z"
                },
                "deletedAt": {
                    "type": "string",
                    "example": "1989-07-30T10:05:53Z"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "e4a3b2c1-3b2c-4a3b-a2c1-1b2c3a4b5c6d"
                },
                "price": {
                    "type": "number"
                },
                "serviceName": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "1986-08-02T22:27:19Z"
                }
            }
        },
        "api_structures.ServiceEdit": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "1960-11-10T04:48:09Z"
                },
                "deletedAt": {
                    "type": "string",
                    "example": "1989-07-30T10:05:53Z"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "e4a3b2c1-3b2c-4a3b-a2c1-1b2c3a4b5c6d"
                },
                "price": {
                    "type": "number"
                },
                "serviceName": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "1986-08-02T22:27:19Z"
                }
            }
        },
        "api_structures.ServiceFilter": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "1960-11-10T04:48:09Z"
                },
                "deletedAt": {
                    "type": "string",
                    "example": "1989-07-30T10:05:53Z"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "e4a3b2c1-3b2c-4a3b-a2c1-1b2c3a4b5c6d"
                },
                "price": {
                    "type": "number"
                },
                "serviceName": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "1986-08-02T22:27:19Z"
                }
            }
        }
    }
}